+++ OwlOSC/OwlOSC/OscMessage.cs
@@ -7,8 +7,8 @@ namespace OwlOSC
 {
 	public class OscMessage : OscPacket
 	{
-		public string Address;
-		public List<object> Arguments;
+		public string Address {get;private set;}
+		public List<object> Arguments {get;private set;}
 
 		public OscMessage(string address, params object[] args)
 		{
@@ -17,6 +17,10 @@ namespace OwlOSC
 			Arguments.AddRange(args);
 		}
 
+		/// <summary>
+		/// Get raw byte data of message for Send purpouse
+		/// </summary>
+		/// <returns>Raw Byte Array</returns>
 		public override byte[] GetBytes()
 		{
 			List<byte[]> parts = new List<byte[]>();
@@ -64,7 +68,7 @@ namespace OwlOSC
 						typeString += "t";
 						parts.Add(setULong((UInt64)arg));
 						break;
-					case "SharpOSC.Timetag":
+					case "OwlOSC.Timetag":
 						typeString += "t";
 						parts.Add(setULong(((Timetag)arg).Tag));
 						break;
@@ -80,7 +84,7 @@ namespace OwlOSC
 						}
 						break;
 
-					case "SharpOSC.Symbol":
+					case "OwlOSC.Symbol":
 						typeString += "S";
 						parts.Add(setString(((Symbol)arg).Value));
 						break;
@@ -89,11 +93,11 @@ namespace OwlOSC
 						typeString += "c";
 						parts.Add(setChar((char)arg));
 						break;
-					case "SharpOSC.RGBA":
+					case "OwlOSC.RGBA":
 						typeString += "r";
 						parts.Add(setRGBA((RGBA)arg));
 						break;
-					case "SharpOSC.Midi":
+					case "OwlOSC.Midi":
 						typeString += "m";
 						parts.Add(setMidi((Midi)arg));
 						break;
@@ -157,8 +161,116 @@ namespace OwlOSC
 			return output;
 		}
 
+		private string FormatArgumentsType(){
+			List<object> currentList = Arguments;
+			int ArgumentsIndex = 0;
+
+			string typeString = "";
+			int i = 0; 
+			while (i < currentList.Count)
+			{
+				var arg = currentList[i];
+
+				string type = (arg != null) ? arg.GetType().ToString() : "null";
+				switch (type)
+				{
+					case "System.Int32":
+						typeString += "i " + arg.ToString();
+						break;
+					case "System.Single":
+						if (float.IsPositiveInfinity((float)arg))
+						{
+							typeString += "I";
+						}
+						else
+						{
+							typeString += "f " + arg.ToString();
+						}
+						break;
+					case "System.String":
+						typeString += "s " + arg.ToString();
+						break;
+					case "System.Byte[]":
+						typeString += $"b (System.Byte[{((byte[])arg).Length}])";
+						break;
+					case "System.Int64":
+						typeString += "h " + arg.ToString();
+						break;
+					case "System.UInt64":
+						typeString += "t " + arg.ToString();
+						break;
+					case "OwlOSC.Timetag":
+						typeString += "t "  + arg.ToString();
+						break;
+					case "System.Double":
+						if (Double.IsPositiveInfinity((double)arg))
+						{
+							typeString += "I";
+						}
+						else
+						{
+							typeString += "d " + arg.ToString();
+						}
+						break;
+
+					case "OwlOSC.Symbol":
+						typeString += "S " + arg.ToString();
+						break;
+
+					case "System.Char":
+						typeString += "c " + arg.ToString();
+						break;
+					case "OwlOSC.RGBA":
+						typeString += "r " + arg.ToString();
+						break;
+					case "OwlOSC.Midi":
+						typeString += "m " + arg.ToString();
+						break;
+					case "System.Boolean":
+						typeString += ((bool)arg) ? "T" : "F";
+						break;
+					case "null":
+						typeString += "N";
+						break;
+
+					// This part handles arrays. It points currentList to the array and resets i
+					// The array is processed like normal and when it is finished we replace  
+					// currentList back with Arguments and continue from where we left off
+					case "System.Object[]":
+					case "System.Collections.Generic.List`1[System.Object]":
+						if(arg.GetType() == typeof(object[]))
+							arg = ((object[])arg).ToList();
+
+						if (Arguments != currentList)
+							throw new Exception("Nested Arrays are not supported");
+						typeString += "[";
+						currentList = (List<object>)arg;
+						ArgumentsIndex = i;
+						i = 0;
+						continue;
+
+					default:
+						throw new Exception("Unable to transmit values of type " + type);
+				}
+
+				i++;
+				if (currentList != Arguments && i == currentList.Count)
+				{ 
+					// End of array, go back to main Argument list
+					typeString += "]";
+					currentList = Arguments;
+					i = ArgumentsIndex+1;
+				}
+
+				if(i < currentList.Count)
+					typeString += ", ";
+			}
+			return typeString;
+		}
+
 		public override string ToString(){
-			return $"'{this.Address}' -> {string.Join(";",this.Arguments)}";
+			//return $"'{this.Address}' -> {string.Join(";",this.Arguments)}";
+			return $"'{this.Address}' -> " + FormatArgumentsType();
 		}
 	}
 }