+++ OwlOSC/OwlOSC/UDPSender.cs
@@ -58,7 +58,8 @@ namespace OwlOSC
 		{
 			if(message.Length > _MAX_PACKET_SIZE)
 				throw new Exception("Message exceeds UDP max packet size (64k)");
-			sock.SendTo(message, RemoteIpEndPoint);
+			int sent = sock.SendTo(message, RemoteIpEndPoint);
+			Console.WriteLine($"Sent {sent} byte of data)");
 		}
 
 		public void Send(OscPacket packet)